- Update memset and memcpy in libc to take 32 bit params as 3rd parameter (We had added modifications that make the C pointer 8 bytes. We need the following changes to keep the expectations of C correct) - Pass the --wasm-host-triple flag to the wasm_clang compiler to ensure we compile with LP64 - Change the cached string for precompoled libraries so that we rebuild the standard library in a system with emscripten already installed

+++ system/include/libc/sched.h
@@ -70,7 +70,7 @@ int clone (int (*)(void *), void *, int, void *, ...);
 int unshare(int);
 int setns(int, int);
 
-void *memcpy(void *__restrict, const void *__restrict, size_t);
+void *memcpy(void *__restrict, const void *__restrict, unsigned int);
 int memcmp(const void *, const void *, size_t);
 void *calloc(size_t, size_t);
 void free(void *);
