+++ system/lib/libc/musl/src/string/memcpy.c
@@ -2,7 +2,7 @@
 #include <stdint.h>
 #include <endian.h>
 
-void *memcpy(void *restrict dest, const void *restrict src, size_t n)
+void *memcpy(void *restrict dest, const void *restrict src, unsigned int n)
 {
 	unsigned char *d = dest;
 	const unsigned char *s = src;
